Sample App to demonstrate Authentication (#270)
* Sample App to demostrate Authentication (Freshchat)

This is sample app to demonstrate Freshchat authentication. It authenticates the user and displays the list of all agents configured by the business that uses Freshchat.

* Sample App to demonstrate Authentication (Freshchat)

This is sample app to demonstrate Freshchat authentication. It authenticates the user and displays the list of all agents configured by the business that uses Freshchat.

* Added screenshots and updated styles

* Removed jQuery and other dependencies that relied on jQuery  (#269)

* Removed jQuery and other dependencies that relied on jQuery such as Boostrap (previously used to align one element), and the select2 library which also had jquery as a dependency but was not used anywhere in the project. Converted relevant jQuery code to Vanilla ES5 javascript equivalents.

* Added in newline characters at the end of js and css files

* IDE formatter seemed to be deleting my newline char at the end of the files, fixed now

* Sample App to demostrate Authentication (Freshchat)

This is sample app to demonstrate Freshchat authentication. It authenticates the user and displays the list of all agents configured by the business that uses Freshchat.

* Sample App to demonstrate Authentication (Freshchat)

This is sample app to demonstrate Freshchat authentication. It authenticates the user and displays the list of all agents configured by the business that uses Freshchat.

* Added screenshots and updated styles

* revised code

Co-authored-by: Ethan Soo Hon <ethanny2@gmail.com>